+++ TeamProject/static/js/post/event.js
@@ -258,20 +258,6 @@ export function handle_report() {
 //==========댓글 crud===========//
 //재민 part
 export function handle_commentInsert() {
-  // const token = sessionStorage.getItem('access_token');
-  // if (token === null) {
-  //   alert('로그인을 먼저 해주세요');
-  //   return null;
-  // }
-
-  // const post_id = event.currentTarget.id.split('_')[2];
-  // await input_comment(post_id);
-  // const footer = document.querySelector('.footer').offsetTop;
-  // window.scrollTo({
-  //   top: footer,
-  //   behavior: 'smooth'
-  // });
-
   const commentBtn = document.querySelector('.btn_comment_input');
   //파일 제출 버튼 태그
   commentBtn.addEventListener("click", async function() { // 제출 이벤트 리스너
@@ -292,30 +278,37 @@ export function handle_commentInsert() {
 //==========댓글 삭제 이벤트==========
 //재민part
 export function handle_commentDelete() {
-
   const confirmflag = confirm("삭제하시겠습니까?");
   const comment_id = event.currentTarget.id.split('__')[1];
   if (confirmflag) MAIN.delete_comment(comment_id);
 }
+
 //==========댓글 수정 이벤트==========
 //재민part
 export function handle_commentUpdate() {
-  console.log("수정창");
-  const comment_id = event.currentTarget.id.split('__')[1];
-  MAIN.update_comment(comment_id);
+  const commentUpdateBtn = document.querySelector('.btn_comment_update');
+  commentUpdateBtn.addEventListener('click', async function(){
+    const comment_id = event.currentTarget.id.split('__')[1];
+    MAIN.update_comment(comment_id);
+  });
 }
+
 //==========댓글 수정 후 완료 이벤트==========
 //재민part
 export const handle_commnetUpdateSubmit = () => {
-  const comment_id = event.currentTarget.id.split('__')[1];
-  MAIN.update_commentSubmit(comment_id);
+  const commentUpdateSubmitBtn = document.querySelector('.btn_comment_update_submit');
+  commentUpdateSubmitBtn.addEventListener('click', async function(){
+    const comment_id = event.currentTarget.id.split('__')[1];
+    MAIN.update_commentSubmit(comment_id);
+  }); 
 }
+
 //==========댓글 좋아요 이벤트===========//
 //재민 part
 export function handle_Commentlikes() {
-  const commentlikeBtn = document.querySelectorAll(".btn_comment_likes");
-  for(let i=0;i<commentlikeBtn.length;i++){
-    commentlikeBtn[i].addEventListener('click', async function() {
+  const commentLikeBtn = document.querySelectorAll(".btn_comment_likes");
+  for(let i=0;i<commentLikeBtn.length;i++){
+    commentLikeBtn[i].addEventListener('click', async function() {
       const token = sessionStorage.getItem('access_token');
       if (token === null) {
         alert('로그인을 먼저 해주세요');
@@ -335,28 +328,33 @@ export function handle_Commentlikes() {
       }
     });
   }
-  
 }
+
 //========== 댓글 신고 이벤트===========//
 //재민 part
-export async function handle_commentReport() {
-  const token = sessionStorage.getItem('access_token');
-  if (token === null) {
-    alert('로그인을 먼저 해주세요');
-    return null;
-  }
-  const target = event.currentTarget;
-  const comment_id = target.id.split('_')[2];
-
-  const check = await MAIN.add_report('comment', comment_id);
-  if (check == true) {
-    alert('신고가 접수 되었습니다.')
-  } else if (check == 403) { //자신의 글일때
-    alert('유효하지 않은 토큰입니다. ');
-  } else if (check == 409) { //이미추천한글일때
-    alert('이미 신고한 글입니다.');
+export function handle_commentReport() {
+  const commentReportBtn = document.querySelectorAll('.btn_comment_report');
+  for(let i=0;i<commentReportBtn.length;i++){
+    commentReportBtn[i].addEventListener('click', async function(){
+    const token = sessionStorage.getItem('access_token');
+    if (token === null) {
+      alert('로그인을 먼저 해주세요');
+      return null;
+    }
+    const target = event.currentTarget;
+    const comment_id = target.id.split('_')[2];
+    const check = await MAIN.add_report('comment', comment_id);
+    if (check == true) {
+      alert('신고가 접수 되었습니다.')
+    } else if (check == 403) { //자신의 글일때
+      alert('유효하지 않은 토큰입니다. ');
+    } else if (check == 409) { //이미추천한글일때
+      alert('이미 신고한 글입니다.');
+    }
+    });
   }
 }
+
 //==========top 버튼 ===========//
 (function handle_goTop() {
   const ele = document.querySelector('.post_goTop');
