+++ TeamProject/templates/main.html
@@ -105,8 +105,8 @@
         AOS.init();
     </script>
     <script src="../static/js/main.js" type="text/javascript"></script>
-    <script src="../static/js/auth.js" type="text/javascript"></script>
-
+    <!-- <script src="../static/js/auth.js" type="text/javascript"></script> -->
+    <script src="../static/js/auth.js" type="module"></script>
 </body>
